# polybot/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Polymarket Trading Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bot/control:
    post:
      operationId: controlBot
      tags: [bot]
      summary: Start, stop or pause the bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotControlInput"
      responses:
        "200":
          description: Updated status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotStatus"
  /bot/config:
    get:
      operationId: getBotConfig
      tags: [bot]
      summary: Get bot configuration
      responses:
        "200":
          description: Bot config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotConfig"
    put:
      operationId: updateBotConfig
      tags: [bot]
      summary: Update bot configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotConfigInput"
      responses:
        "200":
          description: Updated config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotConfig"
  /markets:
    get:
      operationId: getMarkets
      tags: [markets]
      summary: List scanned markets
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, resolved, all]
            default: active
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Markets list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Market"
  /markets/{marketId}:
    get:
      operationId: getMarket
      tags: [markets]
      summary: Get single market details
      parameters:
        - name: marketId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Market detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Market"
...
  /positions/{positionId}:
    get:
      operationId: getPosition
      tags: [positions]
      summary: Get position details
      parameters:
        - name: positionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Position
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
...
  /pnl/chart:
    get:
      operationId: getPnlChart
      tags: [pnl]
      summary: PnL chart data points over time
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [1d, 7d, 30d, all]
            default: 7d
      responses:
        "200":
          description: PnL chart data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PnlDataPoint"
...
  /paper/positions:
    get:
      operationId: getPaperPositions
      tags: [paper]
      summary: List paper trading positions
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, closed, all]
            default: open
      responses:
        "200":
          description: Paper positions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaperPosition"
  /paper/trades:
    get:
      operationId: getPaperTrades
      tags: [paper]
      summary: List paper trading history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: Paper trades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaperTrade"
...
  /paper/chart:
    get:
      operationId: getPaperChart
      tags: [paper]
      summary: Paper portfolio equity curve
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [1d, 7d, 30d, all]
            default: 7d
      responses:
        "200":
          description: Paper chart data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaperDataPoint"
...
components:
  schemas:
...
    BotStatus:
      type: object
      properties:
        state:
          type: string
          enum: [running, stopped, paused, error]
        uptime_seconds:
          type: ["integer", "null"]
        markets_scanned:
          type: integer
        last_scan_at:
          type: ["string", "null"]
        error_message:
          type: ["string", "null"]
        pid:
          type: ["integer", "null"]
      required: [state, markets_scanned]
    BotControlInput:
      type: object
      properties:
        action:
          type: string
          enum: [start, stop, pause, resume]
      required: [action]
    BotConfig:
      type: object
      properties:
        id:
          type: integer
        edge_threshold:
          type: number
          description: Minimum edge % to trigger a signal (e.g. 0.05 = 5%)
        max_position_pct:
          type: number
          description: Max % of capital per market
        daily_loss_limit_pct:
          type: number
        kelly_fraction:
          type: number
          description: Kelly criterion fraction (e.g. 0.25 for quarter-Kelly)
        scan_interval_seconds:
          type: integer
        use_limit_orders:
          type: boolean
        min_liquidity_usd:
          type: number
        max_correlated_exposure_pct:
          type: number
        paper_trading:
          type: boolean
          description: Enable paper trading mode (virtual money, real prices)
        paper_capital_usd:
          type: number
          description: Virtual capital for paper trading in USD
        updated_at:
          type: string
      required:
        - id
        - edge_threshold
        - max_position_pct
        - daily_loss_limit_pct
        - kelly_fraction
        - scan_interval_seconds
        - use_limit_orders
        - min_liquidity_usd
        - max_correlated_exposure_pct
        - paper_trading
        - paper_capital_usd
    BotConfigInput:
      type: object
      properties:
        edge_threshold:
          type: number
        max_position_pct:
          type: number
        daily_loss_limit_pct:
          type: number
        kelly_fraction:
          type: number
        scan_interval_seconds:
          type: integer
        use_limit_orders:
          type: boolean
        min_liquidity_usd:
          type: number
        max_correlated_exposure_pct:
          type: number
        paper_trading:
          type: boolean
        paper_capital_usd:
          type: number
    Market:
      type: object
      properties:
        id:
          type: string
        question:
          type: string
        category:
          type: string
        market_price:
          type: number
          description: Current YES token price (0-1)
        model_probability:
          type: ["number", "null"]
          description: Bot's estimated fair probability
        edge:
          type: ["number", "null"]
          description: Edge = model_probability - market_price
        volume_24h:
          type: number
        liquidity_usd:
          type: number
        end_date:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, resolved, cancelled]
        last_updated:
          type: string
      required: [id, question, market_price, volume_24h, liquidity_usd, status, last_updated]
    Position:
      type: object
      properties:
        id:
          type: integer
        market_id:
          type: string
        market_question:
          type: string
        side:
          type: string
          enum: [YES, NO]
        size_usd:
          type: number
        entry_price:
          type: number
        current_price:
          type: number
        unrealized_pnl:
          type: number
        unrealized_pnl_pct:
          type: number
        entry_edge:
          type: number
        current_edge:
          type: ["number", "null"]
        opened_at:
          type: string
        status:
          type: string
          enum: [open, closed, pending]
      required:
        - id
        - market_id
        - market_question
        - side
        - size_usd
        - entry_price
        - current_price
        - unrealized_pnl
        - unrealized_pnl_pct
        - opened_at
        - status
...
    PnlDataPoint:
      type: object
      properties:
        timestamp:
          type: string
        cumulative_pnl:
          type: number
        portfolio_value:
          type: number
      required: [timestamp, cumulative_pnl, portfolio_value]
...
    PaperPosition:
      type: object
      properties:
        id:
          type: integer
        market_id:
          type: string
        market_question:
          type: string
        side:
          type: string
          enum: [YES, NO]
        size_usd:
          type: number
        entry_price:
          type: number
        current_price:
          type: number
        unrealized_pnl:
          type: number
        unrealized_pnl_pct:
          type: number
        entry_edge:
          type: number
        signal_type:
          type: string
        confidence:
          type: number
        status:
          type: string
        opened_at:
          type: string
        closed_at:
          type: ["string", "null"]
      required:
        - id
        - market_id
        - market_question
        - side
        - size_usd
        - entry_price
        - current_price
        - unrealized_pnl
        - unrealized_pnl_pct
        - entry_edge
        - signal_type
        - confidence
        - status
        - opened_at
    PaperTrade:
      type: object
      properties:
        id:
          type: integer
        market_id:
          type: string
        market_question:
          type: string
        side:
          type: string
          enum: [YES, NO]
        action:
          type: string
          enum: [buy, sell]
        size_usd:
          type: number
        price:
          type: number
        slippage:
          type: number
        fee_usd:
          type: number
        realized_pnl:
          type: ["number", "null"]
        signal_type:
          type: string
        executed_at:
          type: string
      required:
        - id
        - market_id
        - market_question
        - side
        - action
        - size_usd
        - price
        - signal_type
        - executed_at
...
    PaperDataPoint:
      type: object
      properties:
        timestamp:
          type: string
        cumulative_pnl:
          type: number
        portfolio_value:
          type: number
        win_rate:
          type: number
        total_trades:
          type: integer
      required: [timestamp, cumulative_pnl, portfolio_value]
...